requirements.txt
pymongo==3.10.1
azure-storage-blob==12.9.0